+++ app/frontend/src/features/Home.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { Link } from "react-router-dom";
 
-import { newGuideRoute, newPlaceRoute } from "../AppRoutes";
 import Button from "../components/Button";
 import TextBody from "../components/TextBody";
+import { newGuideRoute, newPlaceRoute } from "../routes";
 import useCurrentUser from "./userQueries/useCurrentUser";
 
 export default function Home() {
