+++ app/frontend/src/features/auth/AuthProvider.tsx
@@ -1,7 +1,7 @@
 import React, { Context, ReactNode, useContext, useEffect } from "react";
 import { useHistory } from "react-router-dom";
 
-import { loginRoute } from "../../AppRoutes";
+import { loginRoute } from "../../routes";
 import { setUnauthenticatedErrorHandler } from "../../service/client";
 import useAuthStore, { AuthStoreType } from "./useAuthStore";
