+++ app/frontend/src/features/auth/Logout.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import { useQueryClient } from "react-query";
 import { Redirect } from "react-router-dom";
 
-import { loginRoute } from "../../AppRoutes";
+import { loginRoute } from "../../routes";
 import { useAuthContext } from "./AuthProvider";
 
 export default function Logout() {
