+++ app/frontend/src/features/auth/signup/CompleteSignupForm.tsx
@@ -3,12 +3,12 @@ import React, { useEffect, useState } from "react";
 import { Controller, useForm } from "react-hook-form";
 import { useHistory, useLocation, useParams } from "react-router-dom";
 
-import { signupRoute } from "../../../AppRoutes";
 import Autocomplete from "../../../components/Autocomplete";
 import Button from "../../../components/Button";
 import CircularProgress from "../../../components/CircularProgress";
 import TextField from "../../../components/TextField";
 import { HostingStatus } from "../../../pb/api_pb";
+import { signupRoute } from "../../../routes";
 import { service } from "../../../service";
 import {
   nameValidationPattern,
