+++ app/frontend/src/features/communities/CommunityPage.tsx
@@ -2,13 +2,6 @@ import { Breadcrumbs } from "@material-ui/core";
 import React, { useEffect, useState } from "react";
 import { Link, useHistory, useParams } from "react-router-dom";
 
-import {
-  routeToCommunity,
-  routeToDiscussion,
-  routeToGroup,
-  routeToGuide,
-  routeToPlace,
-} from "../../AppRoutes";
 import Alert from "../../components/Alert";
 import CircularProgress from "../../components/CircularProgress";
 import CommentBox from "../../components/Comments/CommentBox";
@@ -19,6 +12,13 @@ import { Community } from "../../pb/communities_pb";
 import { Discussion } from "../../pb/discussions_pb";
 import { Group } from "../../pb/groups_pb";
 import { Page } from "../../pb/pages_pb";
+import {
+  routeToCommunity,
+  routeToDiscussion,
+  routeToGroup,
+  routeToGuide,
+  routeToPlace,
+} from "../../routes";
 import { service } from "../../service";
 
 export default function CommunityPage() {
