+++ app/frontend/src/features/communities/DiscussionPage.tsx
@@ -1,7 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { useHistory, useParams } from "react-router-dom";
 
-import { routeToDiscussion } from "../../AppRoutes";
 import Alert from "../../components/Alert";
 import CircularProgress from "../../components/CircularProgress";
 import CommentBox from "../../components/Comments/CommentBox";
@@ -9,6 +8,7 @@ import Markdown from "../../components/Markdown";
 import PageTitle from "../../components/PageTitle";
 import TextBody from "../../components/TextBody";
 import { Discussion } from "../../pb/discussions_pb";
+import { routeToDiscussion } from "../../routes";
 import { service } from "../../service";
 
 export default function DiscussionPage() {
