+++ app/frontend/src/features/messages/surfing/NewHostRequest.tsx
@@ -7,11 +7,11 @@ import { useForm } from "react-hook-form";
 import { useMutation } from "react-query";
 import { useHistory, useParams } from "react-router-dom";
 
-import { routeToHostRequest } from "../../../AppRoutes";
 import Alert from "../../../components/Alert";
 import Button from "../../../components/Button";
 import TextField from "../../../components/TextField";
 import { CreateHostRequestReq } from "../../../pb/requests_pb";
+import { routeToHostRequest } from "../../../routes";
 import { service } from "../../../service";
 import { firstName } from "../../../utils/names";
 import { validateFutureDate } from "../../../utils/validation";
