+++ app/frontend/src/features/messages/surfing/SurfingTab.tsx
@@ -4,12 +4,12 @@ import * as React from "react";
 import { useQuery } from "react-query";
 import { Link } from "react-router-dom";
 
-import { routeToHostRequest } from "../../../AppRoutes";
 import Alert from "../../../components/Alert";
 import CircularProgress from "../../../components/CircularProgress";
 import TextBody from "../../../components/TextBody";
 import { GroupChat } from "../../../pb/conversations_pb";
 import { HostRequest } from "../../../pb/requests_pb";
+import { routeToHostRequest } from "../../../routes";
 import { service } from "../../../service";
 import useMessageListStyles from "../useMessageListStyles";
 import HostRequestListItem from "./HostRequestListItem";
