+++ app/frontend/src/features/userPage/UserPage.tsx
@@ -2,13 +2,13 @@ import { Grid, makeStyles } from "@material-ui/core";
 import React from "react";
 import { Link, useParams } from "react-router-dom";
 
-import { profileRoute, routeToNewHostRequest } from "../../AppRoutes";
 import Alert from "../../components/Alert";
 import Button from "../../components/Button";
 import CircularProgress from "../../components/CircularProgress";
 import { CouchIcon, EditIcon } from "../../components/Icons";
 import TextBody from "../../components/TextBody";
 import { User } from "../../pb/api_pb";
+import { profileRoute, routeToNewHostRequest } from "../../routes";
 import { useIsMounted, useSafeState } from "../../utils/hooks";
 import AddFriendButton from "../connections/friends/AddFriendButton";
 import useCurrentUser from "../userQueries/useCurrentUser";
